requirements.txt
Django==1.9
djangorestframework==3.3.2
httpie==0.9.3
Pygments==2.1.3
requests==2.9.1
wheel==0.29.0
